chore: rename /utilities to /utils
