+++ public/sitemap.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:mobile="http://www.google.com/schemas/sitemap-mobile/1.0" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
-<url><loc>https://brewww.studio</loc><lastmod>2024-10-23T19:27:35.964Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://brewww.studio/proposal</loc><lastmod>2024-10-23T19:27:35.964Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://brewww.studio</loc><lastmod>2024-10-29T18:46:20.005Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://brewww.studio/proposal</loc><lastmod>2024-10-29T18:46:20.006Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
 </urlset>
