+++ src/components/Button/index.tsx
@@ -2,7 +2,7 @@
 
 import React, { forwardRef } from "react";
 import { cva, type VariantProps } from "class-variance-authority";
-import { cn } from "@/utilities/cn";
+import { cn } from "@/utils/cn";
 import Link from "next/link";
 import { LinkType, Reference } from "../CMSLink/index.js";
 import { Page as PayloadPage } from "@/payload-types";
