+++ src/components/CoolButton/index.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/utilities/cn";
+import { cn } from "@root/utils/cn";
 import { cva, type VariantProps } from "class-variance-authority";
 
 const buttonVariants = cva(
