+++ src/components/UI/Checkbox/index.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { cn } from "@/utilities/cn";
+import { cn } from "@root/utils/cn";
 import * as CheckboxPrimitive from "@radix-ui/react-checkbox";
 import { Check } from "lucide-react";
 import * as React from "react";
