+++ src/components/UI/Select/index.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { cn } from "@/utilities/cn";
+import { cn } from "@root/utils/cn";
 import * as SelectPrimitive from "@radix-ui/react-select";
 import { Check, ChevronDown, ChevronUp } from "lucide-react";
 import * as React from "react";
