+++ src/fields/link.ts
@@ -1,6 +1,6 @@
 import type { Field } from "payload";
 
-import deepMerge from "@/utilities/deepMerge";
+import deepMerge from "@root/utils/deepMerge";
 
 export type LinkAppearances = "default" | "outline";
