+++ src/utils/canUseDOM.ts
