+++ src/utils/formatPagePath.ts
