+++ src/utils/generatePreviewPath.ts
