+++ tsconfig.json
@@ -37,7 +37,7 @@
       "@root/*": ["src/*"],
       "@/styles/*": ["src/styles/*"],
       "@/payload-types": ["./src/payload-types.ts"],
-      "@/utilities/*": ["src/utilities/*"]
+      "@/utils/*": ["src/utils/*"]
     },
     "target": "ES2017"
   },
